Fix case insensitive search issue.
The field 'hashtagWithoutHash' is converted to lowercase when being indexed by the graph. So, we'll always have an all-lowercase hashtag to search against. The Graph doesn't have case-insensitive partial match querying implemented, so since we already are saving the lowercase formatted version of each hashtag, we just need to make sure we're converting our query string to lowercase.